requirements.txt
Django==1.11.13
numpy==1.14.3
djangorestframework==3.8.2
mysqlclient==1.3.6
